993: Fixed the pipeines project layout so that it will load in Eclipse
The source direcory was not working in the 2019.09 version of Eclipse. The
problem was in the layout of the project: after the src/main/java and
src/test/java, the directory was naed com.iluwatar.pipeline. It should've
been com/iluwatar/pipeline. This follows the hierachy of all of the other
patterns.

Once these files were moved, the Pipeline project compiled without errors.